+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This CHANGELOG follows the format listed at [Keep A Changelog](http://keepachang
 - Implemented check for reserved instances
 - Added check to ensure that some or all AWS ConfigService rules have full compliance
 - Added check to ensure that sns subscriptions is not pending
+- handler-ec2_node.rb: protect from instance state_reason which may be nil
 
 ## [2.3.0] - 2016-03-18
 ### Added